Move files into service subdirs, fix imports
