+++ beaconChain/package.json
@@ -18,7 +18,7 @@
     "lint": "eslint --ext .ts src/",
     "lint-fix": "eslint --ext .ts src/ --fix",
     "pretest": "npm run check-types",
-    "test": "nyc -r lcov -e .ts -x \"*.test.ts\" mocha -r ./.babel-register tests/**/*.test.ts && nyc report",
+    "test": "nyc -r lcov -e .ts -x \"*.test.ts\" mocha -r ./.babel-register \"tests/**/*.test.ts\" && nyc report",
     "coverage": "codecov"
   },
   "repository": {
