+++ beaconChain/src/chain/state/index.ts
@@ -1,28 +1,20 @@
 import BN from "bn.js";
-import { hashTreeRoot } from "@chainsafesystems/ssz";
+import {hashTreeRoot} from "@chainsafesystems/ssz";
 
 import {
   BeaconState,
   Crosslink,
   Deposit,
   Eth1Data,
-  int,
+  uint64,
   ValidatorIndex,
-} from "../types";
+} from "../../types";
 
 import {
-  GENESIS_EPOCH,
-  GENESIS_FORK_VERSION,
-  GENESIS_SLOT,
-  GENESIS_START_SHARD,
-  LATEST_ACTIVE_INDEX_ROOTS_LENGTH,
-  LATEST_BLOCK_ROOTS_LENGTH,
-  LATEST_RANDAO_MIXES_LENGTH,
-  LATEST_SLASHED_EXIT_LENGTH,
-  MAX_DEPOSIT_AMOUNT,
-  SHARD_COUNT,
-  ZERO_HASH,
-} from "../constants";
+  GENESIS_EPOCH, GENESIS_FORK_VERSION, GENESIS_SLOT, GENESIS_START_SHARD,
+  LATEST_ACTIVE_INDEX_ROOTS_LENGTH, LATEST_BLOCK_ROOTS_LENGTH, LATEST_RANDAO_MIXES_LENGTH,
+  LATEST_SLASHED_EXIT_LENGTH, MAX_DEPOSIT_AMOUNT, SHARD_COUNT, ZERO_HASH,
+} from "../../constants";
 
 import {
   generateSeed,
@@ -35,6 +27,7 @@ import {
   activateValidator,
 } from "../helpers/validatorStatus";
 
+
 /**
  * Generate the initial beacon chain state.
  * @param {Deposit[]} initialValidatorDeposits
@@ -43,8 +36,8 @@ import {
  * @returns {BeaconState}
  */
 export function getInitialBeaconState(
-  genesisValidatorDeposits: Deposit[],
-  genesisTime: int,
+  initialValidatorDeposits: Deposit[],
+  genesisTime: uint64,
   latestEth1Data: Eth1Data): BeaconState {
 
   const initialCrosslinkRecord: Crosslink = {
@@ -55,7 +48,7 @@ export function getInitialBeaconState(
   const state: BeaconState = {
     // MISC
     slot: GENESIS_SLOT,
-    genesisTime: new BN(genesisTime),
+    genesisTime,
     fork: {
       previousVersion: GENESIS_FORK_VERSION,
       currentVersion: GENESIS_FORK_VERSION,
@@ -92,11 +85,11 @@ export function getInitialBeaconState(
     // PoW receipt root
     latestEth1Data,
     eth1DataVotes: [],
-    depositIndex: new BN(genesisValidatorDeposits.length),
+    depositIndex: new BN(0),
   };
 
   // Process initial deposists
-  genesisValidatorDeposits.forEach((deposit) => {
+  initialValidatorDeposits.forEach((deposit) => {
     processDeposit(
       state,
       deposit.depositData.depositInput.pubkey,
@@ -106,16 +99,16 @@ export function getInitialBeaconState(
     );
   });
 
-  // Process genesis activations
-  // TODO For loop is stubbed, should not cast new BN(i)
-  for (let i = 0; i < state.validatorRegistry.length; i ++) {
-    if (getEffectiveBalance(state, new BN(i)) >= MAX_DEPOSIT_AMOUNT) {
-      activateValidator(state, new BN(i), true);
+  // Process initial activations
+  for (let i: ValidatorIndex = new BN(0); i.ltn(state.validatorRegistry.length); i = i.add(new BN(1))) {
+    // TODO: Unsafe usage of toNumber on i
+    if (getEffectiveBalance(state, i) >= MAX_DEPOSIT_AMOUNT) {
+      activateValidator(state, i, true);
     }
   }
 
   const genesisActiveIndexRoot = hashTreeRoot(getActiveValidatorIndices(state.validatorRegistry, GENESIS_EPOCH), [ValidatorIndex]);
-  for (let index = 0; index < LATEST_ACTIVE_INDEX_ROOTS_LENGTH; index++) {
+  for (let index: number; index < LATEST_ACTIVE_INDEX_ROOTS_LENGTH; index++) {
     state.latestActiveIndexRoots[index] = genesisActiveIndexRoot;
   }
   state.currentShufflingSeed = generateSeed(state, GENESIS_EPOCH);
