+++ beaconChain/src/chain/stateTransition/block/proposerSignature.ts
@@ -7,13 +7,13 @@ import {
   BeaconState,
   bytes32,
   ProposalSignedData,
-} from "../../types";
+} from "../../../types";
 
 import {
   BEACON_CHAIN_SHARD_NUMBER,
   Domain,
   EMPTY_SIGNATURE,
-} from "../../constants";
+} from "../../../constants";
 
 import {
   getBeaconProposerIndex,
@@ -21,7 +21,7 @@ import {
   getDomain,
 } from "../../helpers/stateTransitionHelpers";
 
-import {blsVerify} from "../../stubs/bls";
+import {blsVerify} from "../../../stubs/bls";
 
 export default function processProposerSignature(state: BeaconState, block: BeaconBlock): void {
   const b: BeaconBlock = {
