+++ beaconChain/src/eth1/fake.ts
