+++ beaconChain/tests/chain/helpers/stateTransitionHelpers.test.ts
@@ -1,7 +1,7 @@
 import BN from "bn.js";
 import { assert } from "chai";
 
-import { SLOTS_PER_EPOCH, TARGET_COMMITTEE_SIZE } from "../../src/constants";
+import { SLOTS_PER_EPOCH, TARGET_COMMITTEE_SIZE } from "../../../src/constants";
 import {
   clamp,
   getActiveValidatorIndices,
@@ -21,11 +21,11 @@ import {
   merkleRoot,
   slotToEpoch,
   split,
-} from "../../src/helpers/stateTransitionHelpers";
-import {Epoch, Fork, int, Slot, uint64, Validator} from "../../src/types";
-import {generateAttestationData} from "../utils/attestation";
-import {randBetween} from "../utils/misc";
-import {generateValidator} from "../utils/validator";
+} from "../../../src/chain/helpers/stateTransitionHelpers";
+import {Epoch, Fork, int, Slot, uint64, Validator} from "../../../src/types";
+import {generateAttestationData} from "../../utils/attestation";
+import {randBetween} from "../../utils/misc";
+import {generateValidator} from "../../utils/validator";
 
 describe("intToBytes", () => {
   const zeroedArray = (length) => Array.from({length}, () => 0);
